Updated Network, CustomAutoComplete, AbstractToolTip to use new Network calls.
Added a more modular implementation of abstractGet and autocompleteGet to Network.
Adjusted CustomAutoComplete and AbstractToolTip classes accordingly to use the
new functions.

The end goal is to make Network a stupider class. It should only make calls and
return XML results. It takes success/failure functions to handle the data.