Updated Network, added Graph class and made necessary changes.
Everything now uses the standard four Network calls to access the services.

I created a new class to handle graph operations. It was so spread out and
hard to tell what was happening where. Now a lot of the graph operation are
done using the Graph class.

The DrawGraph class still interacts with the environment, but this should
be added to the Graph class. -- That is something for later.