Update AbstractToolTip
Tooltip now reads all of its information from the abstract xml object. It
initially uses the title it receives from the Node as the title.

The behavior on tooltip has been updated to handle loading more effectively. It
will not do anything when a user clicks on the tooltip while the article is loading.

The tooltip is now a Panel object instead of a BorderContainer. This should be
easier to style going forward and it provides a builtin title bar. It is also
easy to add images to should we decide to add a loading gif (or whatever).

The parse abstract function now returns an object of type Abstract to support the
tooltip changes.