Added loading image to tool tips while loading.
Added the GIFPlayer-0.6 library so that we can load and play animated gifs.
The GIFPlayerComponent class should be used to embed the images. Support for
embeded and non-embeded images is included. Embed images will have to use the
byteArray getter/setter. Non-embeded use url.

Add EmbedAssets so that images we use will be included as part of the SWF.

Updated tool tip class to display loading image while waiting for abstract to arrive.